+++ test/test_regressions.rb
@@ -8,9 +8,11 @@ class TestRegressions < MiniTest::Test
   def setup
     @cached_headers = {
       :cached => {
+        "Age"               => "99999",
         "Date"              => {"httpdate"=>"Thu, 01 Jan 2015 07:03:45 GMT", "timestamp"=>1420095825},
         "Cache-Control"     => {
-          "public"          => {"token"=>nil, "quoted_string"=>nil}
+          "public"          => {"token"=>nil, "quoted_string"=>nil},
+          "max-age"         => {"token"=>"10", "quoted_string"=>nil}
         },
         "X-Cache-Req-Date"  => {"httpdate"=>"Sat, 03 Jan 2015 07:03:45 GMT", "timestamp"=>1420268625},
         "X-Cache-Res-Date"  => {"httpdate"=>"Sat, 03 Jan 2015 07:03:45 GMT", "timestamp"=>1420268625}
@@ -43,5 +45,25 @@ class TestRegressions < MiniTest::Test
 
     assert_equal isnil,   []
     assert_equal isempty, []
+  end
+
+  # https://github.com/aw/CacheRules/issues/10
+  def test_bugfix_10_request_header_max_age_is_checked
+    request_maxage = CacheRules.validate_no_cache?({
+      :cached   => @cached_headers[:cached],
+      :request  => {"Cache-Control" => {"max-age" => {"token"=>0, "quoted_string"=>nil} } }
+    })
+    current = CacheRules.validate_no_cache?({
+      :cached   => @cached_headers[:cached],
+      :request  => {"Cache-Control" => {"max-age" => {"token"=>1000, "quoted_string"=>nil} } }
+    })
+    cached_max_age = CacheRules.validate_expired?({
+      :cached   => @cached_headers[:cached],
+      :request  => {}
+    })
+
+    assert_equal 1, request_maxage
+    assert_equal 1, current
+    assert_equal 1, cached_max_age
   end
 end